% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{add}
\usage{
add(a, b)
}
\arguments{
\item{a}{a number}

\item{b}{another number}
}
\value{
the sum of the two
}
\description{
add
}
\note{
this is from \href{http://adv-r.had.co.nz/C-interface.html}{Hadely's guide on C}
}
\examples{
regolith::add(1, 4)
}
