% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{eb}
\alias{eb}
\title{Number of possible rearrangements}
\usage{
eb(x, rqs = TRUE)
}
\arguments{
\item{x}{an integer vector}

\item{rqs}{a \code{logical} indicating whether or not R's version of quicksort should be used. Defaults to \code{TRUE}.}
}
\value{
a number indicating the number of possible rearrangements for a given
vector using binomial expansion: \code{n!/(a!b!c!)}
}
\description{
Number of possible rearrangements
}
\examples{
x <- c(10L, 9L, 9L, 8L, 10L)
eb(x)
eb(c(1, 2))
eb(c(1, 1))
eb(c(1, 2, 3, 3))
}
