% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_parallel.R
\name{has_parallel}
\alias{has_parallel}
\title{Does thie package have paralell capability via OMP?}
\usage{
has_parallel()
}
\value{
a logical \code{TRUE} or \code{FALSE}
}
\description{
Does thie package have paralell capability via OMP?
}
\examples{
has_parallel()
}
