% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{sorts}
\alias{sorts}
\title{sorts}
\usage{
sorts(x, R = TRUE)
}
\arguments{
\item{R}{if sort should be done by \code{R_qsort()} or \code{qsort()}}
}
\value{
a number
}
\description{
sorts
}
\examples{
sorts(sample(10, replace = TRUE))
}
