% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{sorts}
\alias{sorts}
\title{sorts}
\usage{
sorts(x, R = TRUE, print = FALSE)
}
\arguments{
\item{R}{if sort should be done by \code{R_qsort()} or \code{qsort()}}

\item{print}{if \code{R = FALSE}, should the comparison be printed?}
}
\value{
a number
}
\description{
sorts
}
\examples{
(x <- sample(10, replace = TRUE))
sorts(x)
x # has side-effect
(x <- sample(10, replace = TRUE))
sorts(x, R = FALSE)
sorts(x, R = FALSE, print = TRUE)
}
